requirements.txt
# Core dependencies
PyMuPDF>=1.23.6        # PDF processing
pytesseract>=0.3.10    # OCR
Pillow>=10.1.0         # Image processing
numpy>=1.26.1          # Numerical operations
scikit-learn>=1.3.2    # For vector similarity calculations

# Embedding models
sentence-transformers>=2.2.2  # Text embeddings
transformers>=4.35.2          # CLIP model for image embeddings
torch>=2.6.0                  # Required for transformers

# LLM interface
requests==2.31.0       # API requests
python-dotenv==1.0.0   # Environment variables

# UI
gradio==4.13.0         # Web interface

# Utilities
tqdm==4.66.1           # Progress bars
python-magic==0.4.27   # File type detection